style(matchers,serializer): add a newline to serialize, rename matchers file
